requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alpha_vantage==3.0.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorlog==6.9.0
curl_cffi==0.11.4
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
feedparser==6.0.11
finnhub-python==2.4.24
frozendict==2.4.6
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jellyfish==1.2.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-experimental==0.3.4
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langgraph==0.5.0
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.1
langgraph-sdk==0.1.72
langsmith==0.4.4
markdown-it-py==3.0.0
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.6.2
multitasking==0.0.11
mypy_extensions==1.1.0
networkx==3.4.2
newsapi-python==0.2.7
numpy==2.2.6
oauthlib==3.3.1
openai==1.93.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.0
peewee==3.18.1
platformdirs==4.3.8
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
questionary==2.1.0
reddit-api==0.4
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
segtok==1.5.11
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stockstats==0.6.5
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
tweepy==4.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
update-checker==0.18.0
urllib3==2.5.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
xxhash==3.5.0
yake==0.4.8
yarl==1.20.1
yfinance==0.2.64
zstandard==0.23.0
